refactor(app): Use typescript
Use typescript
