+++ package.json
@@ -3,7 +3,8 @@
         "@rsbuild/core": "^1.1.6",
         "autoprefixer": "^10.4.19",
         "postcss": "^8.4.38",
-        "tailwindcss": "^3.4.4"
+        "tailwindcss": "^3.4.4",
+        "typescript": "^5.7.2"
     },
     "dependencies": {
         "@tweenjs/tween.js": "^23.1.1",
