+++ src/modules/PropDetail.ts
