+++ src/modules/PropDynamic.ts
