+++ src/modules/PropInteractable.ts
