+++ src/modules/PropStatic.ts
