+++ src/props/LightEnvironment.ts
