+++ src/props/PlayerCamera.ts
