+++ tsconfig.json
@@ -1,5 +1,9 @@
 {
   "compilerOptions": {
-    "isolatedModules": true
+    "isolatedModules": true,
+    "lib": [
+      "ES2015",
+      "dom"
+    ]
   }
 }
